+++ CHANGELOG.md
@@ -9,7 +9,7 @@ The format is based on [Keep a Changelog], and this project adheres to
 [Keep a Changelog]: https://keepachangelog.com/en/1.0.0/
 [Semantic Versioning]: https://semver.org/spec/v2.0.0.html
 
-## [Unreleased]
+## [0.7.0] - 2021-05-05
 
 This release upgrades the protocol buffers implementation to use
 `google.golang.org/protobuf` instead of `github.com/golang/protobuf`.
@@ -111,6 +111,7 @@ different marshalers/unmarshalers provided by `google.golang.org/protobuf`.
 [0.4.0]: https://github.com/dogmatiq/marshalkit/releases/tag/v0.4.0
 [0.5.0]: https://github.com/dogmatiq/marshalkit/releases/tag/v0.5.0
 [0.6.0]: https://github.com/dogmatiq/marshalkit/releases/tag/v0.6.0
+[0.7.0]: https://github.com/dogmatiq/marshalkit/releases/tag/v0.7.0
 
 <!-- version template
 ## [0.0.1] - YYYY-MM-DD